refactor: update path aliases from @/shared/* to @/core/*
- Update all tsconfig.json files to use @/core/* alias for core package imports
- Modify vitest configurations to resolve @/core/* paths correctly
- Update ModelManager files to use @/core/* import paths for consistency
- Maintain separate @/* alias for local package imports

This provides clearer distinction between core package imports (@/core/*)
and local package imports (@/*), improving code organization and readability.
