+++ packages/api/tsconfig.json
@@ -3,9 +3,10 @@
   "compilerOptions": {
     "outDir": "./dist",
     "rootDir": "./src",
-    "baseUrl": "./src",
+    "baseUrl": ".",
     "paths": {
-      "@/*": ["./*", "../core/src/*"]
+      "@/*": ["./src/*"],
+      "@/core/*": ["../core/src/*"]
     }
   },
   "include": ["src/**/*"],
