+++ packages/cli/tsconfig.json
@@ -19,9 +19,10 @@
     "noImplicitReturns": false,
     "noFallthroughCasesInSwitch": false,
     "noUncheckedIndexedAccess": false,
-    "baseUrl": "./src",
+    "baseUrl": ".",
     "paths": {
-      "@/*": ["../core/src/*"]
+      "@/*": ["./src/*"],
+      "@/core/*": ["../core/src/*"]
     }
   },
   "include": ["src/**/*"],
