+++ packages/core/src/shared/models/model-manager.ts
@@ -5,14 +5,14 @@
 
 import { count, eq, sql } from "drizzle-orm"
 import { Context, Effect, Layer } from "effect"
-import { DatabaseService } from "@/shared/database/connection"
-import { embeddings } from "@/shared/database/schema"
-import { DatabaseQueryError } from "@/shared/errors/database"
+import { DatabaseService } from "@/core/shared/database/connection"
+import { embeddings } from "@/core/shared/database/schema"
+import { DatabaseQueryError } from "@/core/shared/errors/database"
 import {
   EmbeddingProviderService,
   type EmbeddingRequest,
   type ModelInfo as ProviderModelInfo,
-} from "@/shared/providers"
+} from "@/core/shared/providers"
 import type {
   ModelManagerInfo,
   ModelCompatibility,
