+++ packages/core/tsconfig.json
@@ -21,9 +21,9 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedIndexedAccess": true,
     "verbatimModuleSyntax": false,
-    "baseUrl": "./src",
+    "baseUrl": ".",
     "paths": {
-      "@/*": ["./*"]
+      "@/core/*": ["./src/*"]
     }
   },
   "include": ["src/**/*"],
