+++ packages/core/vitest.config.ts
@@ -5,7 +5,7 @@ import { resolve } from 'path'
 export default defineConfig({
   resolve: {
     alias: {
-      '@': resolve(__dirname, 'src')
+      '@/core': resolve(__dirname, 'src')
     }
   },
   test: {
